make referenced vs unreferenced compositor tests
an unreferenced test passes as long as the operations complete without reporting failure

a referenced test has a reference image which the result of composition must remain identical to

in case the sicgl functions that draw the source images change copies of these images are stored and cross-checked. in the future it would be simpler and faster to simply load these png files straight into a sicgl interface and use that as input to the compositor functions.